chore: add settings tab component
